week3/day3/chatbot/requirements.txt
flask
openai
python-dotenv
python-docx
PyPDF2
pillow
pytesseract
requests
